refactor: move AI field marker to field-highlighter
field-base only provides mixins, while the marker is an element that
decorates a field from the outside, like the highlighter does.
