+++ dev/ai-field-marker.html
@@ -35,7 +35,7 @@
       import '@vaadin/text-area';
       import '@vaadin/text-field';
       import '@vaadin/radio-group';
-      import '@vaadin/field-base/src/vaadin-ai-field-marker.js';
+      import '@vaadin/field-highlighter/src/vaadin-ai-field-marker.js';
 
       const fields = [...document.querySelectorAll('vaadin-form-layout > *')];
