+++ packages/field-base/package.json
@@ -34,18 +34,13 @@
     "@open-wc/dedupe-mixin": "^1.3.0",
     "@vaadin/a11y-base": "25.3.0-alpha8",
     "@vaadin/component-base": "25.3.0-alpha8",
-    "@vaadin/popover": "25.3.0-alpha8",
-    "@vaadin/tooltip": "25.3.0-alpha8",
-    "@vaadin/vaadin-themable-mixin": "25.3.0-alpha8",
     "lit": "^3.0.0"
   },
   "devDependencies": {
     "@vaadin/chai-plugins": "25.3.0-alpha8",
-    "@vaadin/custom-field": "25.3.0-alpha8",
     "@vaadin/input-container": "25.3.0-alpha8",
     "@vaadin/test-runner-commands": "25.3.0-alpha8",
     "@vaadin/testing-helpers": "^2.0.0",
-    "@vaadin/text-field": "25.3.0-alpha8",
     "sinon": "^22.0.0"
   },
   "customElements": "custom-elements.json"
