+++ packages/field-highlighter/package.json
@@ -35,6 +35,8 @@
     "@vaadin/a11y-base": "25.3.0-alpha8",
     "@vaadin/component-base": "25.3.0-alpha8",
     "@vaadin/overlay": "25.3.0-alpha8",
+    "@vaadin/popover": "25.3.0-alpha8",
+    "@vaadin/tooltip": "25.3.0-alpha8",
     "@vaadin/vaadin-themable-mixin": "25.3.0-alpha8",
     "lit": "^3.0.0"
   },
@@ -43,6 +45,7 @@
     "@vaadin/chai-plugins": "25.3.0-alpha8",
     "@vaadin/checkbox": "25.3.0-alpha8",
     "@vaadin/checkbox-group": "25.3.0-alpha8",
+    "@vaadin/custom-field": "25.3.0-alpha8",
     "@vaadin/date-picker": "25.3.0-alpha8",
     "@vaadin/date-time-picker": "25.3.0-alpha8",
     "@vaadin/item": "25.3.0-alpha8",
