+++ packages/field-highlighter/src/styles/vaadin-ai-field-marker-base-styles.js
