+++ packages/field-highlighter/src/vaadin-ai-field-marker.js
