+++ packages/field-highlighter/test/ai-field-marker-property-registration.test.js
