+++ packages/field-highlighter/test/visual/ai-field-marker-not-animated-styles.css
