+++ packages/field-highlighter/test/visual/base/screenshots/ai-field-marker/baseline/ai-marker-working.png
